requirements.txt
Flask==2.1.2
torch==1.12.1
Pillow==9.1.1
ultralytics==0.0.120